separate request_handler to cluster and user request handlers
